requirements.txt
pywebview>=5.4
cryptography>=41.0
pyqt5
pyqtwebengine
